requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.29
python-dotenv==1.0.1
